Fake Programming Language
The following files are used to read files and display output based on
commands entered in a .txt file.

File reader uses a hash table to hold information about certain
variables that get stored when reading the .txt file for output.

The commands to enter in a .txt file are:

1. START-FINISH blocks. Indicating different scopes.
2. COM - Single line comments: Text should be ignored if on the same
line
3. VAR varName – Variable Declaration, adds “varName” to the hash table.
4. variable = expression – Assignment statements, ie GEORGE = 122. Find
GEORGE in the hash table and assign 122 to it.
5. ++ - increment operator, syntax: VARIABLE ++
6. -- - decrement operator, syntax: VARIABLE --
7. expressions, expressions are limited to unary and binary arithmetic,
or variable names
8. supported operators: +  -  /  *  %  ^ (plus, minus, divide, multiple,
modulo, exponent)
9. PRINT – syntax PRINT expression. If the expression is a variable, and
this variable is not in scope, then an error message indicating unknown
variable x at line number y. The value printed if there is a variable in
scope should be the variable with the closest scope.
10. Errors – other than the print statements, our interpreter will not
be responsible for detecting errors, syntax errors should be disregarded
if encountered, assume that the source file is correct.